16.05.2025 - STV: index.html (added buttons), index.js (if-Statement), timeline.js (buttonhandler to save favorites with button in index.html), Server.js (added /fav route), Auth.js (added the favorites to the things that get pushed), favorites.js (use, get post, delete), new file: favorites.html - you can add your own events, delete yours or from the api the ones you saved
it is possible to make Events from the api as favorites and see them in a seperate favorites file where you can only Access with a token, without a token /fav loads Nothing and no possibility to save something (NOTICE: it's only possible as Long as localhost (Server) doesn't get restarted - as soon as something changes (saving document, …) there is a possibility that the token doesn't gets saved. So Restart the Server and than test if you want to use it.